src/app/core/interceptors/http-interceptor.service.spec.ts
import { HTTP_INTERCEPTORS } from '@angular/common/http';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { TestBed } from '@angular/core/testing';
import { environment } from 'src/environments/environment.development';
import { UserService } from '../services/user/user.service';
import { HttpInterceptorService } from './http-interceptor.service';

describe('HttpInterceptorService', () => {
  let service: HttpInterceptorService;
  let httpMock: HttpTestingController;
  let userService: UserService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [
        HttpInterceptorService,
        UserService,
        {
          provide: HTTP_INTERCEPTORS,
          useClass: HttpInterceptorService,
          multi: true
        }
      ]
    });
    service = TestBed.inject(HttpInterceptorService);
    httpMock = TestBed.inject(HttpTestingController);
    userService = TestBed.inject(UserService);
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  it('Should open an alert if there is a error', () => {

    const spy = spyOn(service['toastService'], 'show');
    userService.loginUser({}).subscribe({
      error: (error) => {
        expect(error).toBeTruthy();
      }
    });
    httpMock.expectOne(environment.apiUrl + userService['url']).error(new ProgressEvent('error'))
    expect(spy).toHaveBeenCalled();
  });

});
